refactor(tools): centralize UUID regex and field-condition helpers
