+++ src/pipefy_mcp/tools/field_condition_tools.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import re
 from typing import Any
 
 from mcp.server.fastmcp import FastMCP
@@ -13,7 +12,10 @@ from pipefy_mcp.tools.pipe_config_tool_helpers import (
     build_field_condition_delete_payload,
     build_field_condition_success_payload,
     build_pipe_tool_error_payload,
+    field_condition_actions_error_message,
     handle_pipe_config_tool_graphql_error,
+    normalize_field_condition_actions,
+    strip_expression_ids_for_create,
 )
 from pipefy_mcp.tools.pipe_config_validators import valid_phase_field_id
 
@@ -31,107 +33,6 @@ _CREATE_FIELD_CONDITION_EXTRA_RESERVED = frozenset(
     }
 )
 _UPDATE_FIELD_CONDITION_EXTRA_RESERVED = frozenset({"id"})
-
-_FIELD_CONDITION_PHASE_FIELD_UUID_RE = re.compile(
-    r"^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
-)
-
-
-def field_condition_phase_field_id_looks_like_slug(value: object) -> bool:
-    """True when ``value`` is probably a phase field slug (``id``) instead of ``internal_id``."""
-    if isinstance(value, int):
-        return False
-    if not isinstance(value, str):
-        return False
-    s = value.strip()
-    if not s:
-        return False
-    if _FIELD_CONDITION_PHASE_FIELD_UUID_RE.fullmatch(s):
-        return False
-    if s.isdigit():
-        return False
-    return any(c.isalpha() for c in s)
-
-
-def _normalize_field_condition_actions(
-    actions: list[dict[str, Any]],
-) -> list[dict[str, Any]]:
-    """Shallow-copy each action and map legacy ``hidden`` → ``hide``."""
-    normalized: list[dict[str, Any]] = []
-    for item in actions:
-        row = dict(item)
-        aid = row.get("actionId")
-        if isinstance(aid, str) and aid.strip().lower() == "hidden":
-            row["actionId"] = "hide"
-        normalized.append(row)
-    return normalized
-
-
-def _strip_expression_ids_for_create(condition: dict[str, Any]) -> dict[str, Any]:
-    """Return a copy of ``condition`` with ``id`` removed and nested ints coerced.
-
-    Output keys mirror the input ``condition`` (typically ``expressions``,
-    ``expressions_structure``, plus any other ``ConditionInput`` fields present).
-
-    ``ConditionExpressionInput.id`` is a persisted primary key — sending arbitrary
-    client tokens on create causes ``RECORD_NOT_FOUND``. ``structure_id`` is coerced
-    to ``int`` for consistency (GraphQL ``ID`` scalar).
-    """
-    expressions = condition.get("expressions")
-
-    def _coerce_int(value: Any) -> Any:
-        try:
-            return int(value)
-        except (ValueError, TypeError):
-            return value
-
-    if not isinstance(expressions, list):
-        return condition
-    cleaned: list[dict[str, Any]] = []
-    for expr in expressions:
-        row = {k: v for k, v in expr.items() if k != "id"}
-        sid = row.get("structure_id")
-        if sid is not None:
-            try:
-                row["structure_id"] = int(sid)
-            except (ValueError, TypeError):
-                pass
-        cleaned.append(row)
-    es = condition.get("expressions_structure")
-    if isinstance(es, list):
-        coerced_ints: list[list[Any]] = []
-        for group in es:
-            if isinstance(group, list):
-                coerced_ints.append([_coerce_int(v) for v in group])
-            else:
-                coerced_ints.append([_coerce_int(group)])
-        return {
-            **condition,
-            "expressions": cleaned,
-            "expressions_structure": coerced_ints,
-        }
-    return {**condition, "expressions": cleaned}
-
-
-def _field_condition_actions_error_message(
-    actions: list[dict[str, Any]],
-) -> str | None:
-    if not isinstance(actions, list) or not actions:
-        return "Invalid 'actions': provide a non-empty list of action objects."
-    if not all(isinstance(item, dict) for item in actions):
-        return "Invalid 'actions': each item must be an object/dict."
-    for index, item in enumerate(actions):
-        raw_id = item.get("phaseFieldId")
-        if raw_id is None:
-            continue
-        if field_condition_phase_field_id_looks_like_slug(raw_id):
-            return (
-                f"Invalid actions[{index}] 'phaseFieldId': value looks like a field "
-                "slug (the `id` from get_phase_fields), but Pipefy expects "
-                "`internal_id` from get_phase_fields for field-condition actions. "
-                "See README (Field condition tools)."
-            )
-    return None
 
 
 class FieldConditionTools:
@@ -186,7 +87,7 @@ class FieldConditionTools:
                 return build_pipe_tool_error_payload(
                     message="Invalid 'extra_input': provide an object/dict or omit.",
                 )
-            act_err = _field_condition_actions_error_message(actions)
+            act_err = field_condition_actions_error_message(actions)
             if act_err:
                 return build_pipe_tool_error_payload(message=act_err)
             merged: dict[str, Any] = {
@@ -194,8 +95,8 @@ class FieldConditionTools:
                 for k, v in (extra_input or {}).items()
                 if k not in _CREATE_FIELD_CONDITION_EXTRA_RESERVED
             }
-            condition_for_api = _strip_expression_ids_for_create(condition)
-            actions_for_api = _normalize_field_condition_actions(actions)
+            condition_for_api = strip_expression_ids_for_create(condition)
+            actions_for_api = normalize_field_condition_actions(actions)
             try:
                 raw = await client.create_field_condition(
                     phase_id,
@@ -266,7 +167,7 @@ class FieldConditionTools:
                         ),
                     )
             if actions is not None:
-                act_err = _field_condition_actions_error_message(actions)
+                act_err = field_condition_actions_error_message(actions)
                 if act_err:
                     return build_pipe_tool_error_payload(message=act_err)
 
@@ -276,9 +177,9 @@ class FieldConditionTools:
                 if k not in _UPDATE_FIELD_CONDITION_EXTRA_RESERVED
             }
             if condition is not None:
-                update_attrs["condition"] = _strip_expression_ids_for_create(condition)
+                update_attrs["condition"] = strip_expression_ids_for_create(condition)
             if actions is not None:
-                update_attrs["actions"] = _normalize_field_condition_actions(actions)
+                update_attrs["actions"] = normalize_field_condition_actions(actions)
             if not update_attrs:
                 return build_pipe_tool_error_payload(
                     message=(
